requirements.txt
fastapi==0.110.0
uvicorn[standard]==0.23.1
asyncpg==0.31.0
python-dotenv==1.0.0
pydantic==2.6.2
passlib[bcrypt]==1.7.4
bcrypt==3.2.2
python-jose[cryptography]==3.3.0
